Fixed casting-caused bug

+++ MovieRating.Domain/Service/ReviewService.cs
@@ -90,7 +90,7 @@ namespace MovieRating.Domain.Service
                 }
             }
 
-            averageMovieRate = sumOfRates/totalReviews;
+            averageMovieRate = (double)sumOfRates/totalReviews;
             
             return averageMovieRate;
         }
